whatever

+++ commands/nou.js
@@ -21,6 +21,7 @@ class nou extends Command {
 
     async run(client, message, args) {
         mentionHook.send(`${message.author.username}#${message.author.discriminator} used the **no u** command in the server: ${message.guild.name} (${message.guild.id})`);
+        if (!message.deletable) return;
         message.delete();
         const embed = new Discord.RichEmbed()
             .setColor(`#36393E`)
